@@ -1,16 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
+using System.Linq;
+using System.Threading.Tasks;
 
-namespace InGame.Dtos
+namespace InGame.WebApi.Models
 {
-    public class ProductDto
+    public class Product : Base
     {
-        public int Id { get; set; }
         public int CategoryId { get; set; }
         public string Name { get; set; }
         public byte[] Image { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }        
+        public decimal Price { get; set; }
+        public bool IsActive { get; set; }
     }
 }
