@@ -1,17 +1,18 @@
-﻿using System;
+﻿using Microsoft.AspNetCore.Http;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace InGame.Api.Models
+namespace InGame.WebUi.Models
 {
-    public class Product : Base
-    {        
+    public class ProductVm
+    {
+        public int Id { get; set; }
         public int CategoryId { get; set; }
         public string Name { get; set; }
-        public string ImageUrl { get; set; }
+        public IFormFile Image { get; set; }
         public string Description { get; set; }
         public decimal Price { get; set; }
-        public bool IsActive { get; set; }
     }
 }
