@@ -51,7 +51,7 @@
                         <ul class="kt-menu__nav ">
                             <li class="kt-menu__item " aria-haspopup="true"><a asp-controller="Home" asp-action="Index" class="kt-menu__link "><i class="kt-menu__link-icon flaticon2-protection"></i><span class="kt-menu__link-text">Dashboard</span></a></li>
                             <li class="kt-menu__item " aria-haspopup="true"><a asp-controller="Category" asp-action="Index" class="kt-menu__link "><i class="kt-menu__link-icon flaticon2-medical-records-1"></i><span class="kt-menu__link-text">Categories</span></a></li>
-                            <li class="kt-menu__item " aria-haspopup="true"><a href="#" class="kt-menu__link "><i class="kt-menu__link-icon flaticon2-medical-records-1"></i><span class="kt-menu__link-text">Products</span></a></li>
+                            <li class="kt-menu__item " aria-haspopup="true"><a asp-controller="Product"  asp-action="Index" class="kt-menu__link "><i class="kt-menu__link-icon flaticon2-medical-records-1"></i><span class="kt-menu__link-text">Products</span></a></li>
                             <li class="kt-menu__item " aria-haspopup="true"><a href="#" class="kt-menu__link "><i class="kt-menu__link-icon flaticon2-medical-records-1"></i><span class="kt-menu__link-text">Users</span></a></li>
                         </ul>
                     </div>
